Progress with EM test behaviors
